requirements.txt
setuptools==67.7.2
kafka-python==2.0.2
